backend/requirements.txt
django>=5.0,<6.0
psycopg2-binary>=2.9.9
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.1
django-environ>=0.11.2
django-cors-headers>=4.3.1
drf-spectacular>=0.27.1
django-filter>=23.3
gunicorn>=21.2.0
whitenoise>=6.6.0
